Moved BootstrapNavDropdownListItem to separate file; added props validation
